Add files via upload
I added a full stop in front of the forward slash I used to specify my file path. This is to tell the code to look for my file/directory in the current directory. This will make it easier to locate the file I want to link to and avoid unexpected issues.